Entity extraction system revamped.
Well, I spent too long doing this so i kinda forgot what changed. I'll just put down how it works now.

The program will loop through a bunch of stuff.
It will loop through all the nestedOutlines that contains the phrases to match. (e.g. [what, day, !daysrelative]). GetExtractedEntities() will run on every nestedOutline.

Firstly, the program populates entityRequests, which is used to track what the next request is IN ORDER. (e.g. If there are multiple requests in a outline, [what !day, !daysrelative, !countries], this entityRequest will line them up in order to be extracted.)
Even if the word matches another entity request, it will ignore those, and only match entities that belong to the request.

Then, the program will PopulateDict first, with whatever request it is on right now. (e.g. !days > mon, tues, wed...). This dict will be refreshed on every request.

The program goes down the user's sentence and finds any matches to the dict.

For multiple words in the dict, here's what I did:
entityLengths stores every available iteration of the lengths of each request. e.g. "monday" is 1 word length, "the day after" is 3 word length. So, in !daysrelative, there are phrases with only 1 word length and 3 word lengths. Hence entityLengths will be [1,3].
Using this, I can concatnate 1 and 3 words ahead of the current searched word, and compare it against the dict for a match. It is modular, as i can do as long of a concatnation as i want, and also for however many different lengths of concatnation.

If there is a match, then the matching will be cut off, the dict will be refreshed with the next set of entities (if any), and the thing goes on.

Then the rest is just there, i'm too lazy to explain the rest of it. It just works lol.